Split up App.tsx substantially
